passing all tool test cases

+++ concurrent/verifier/item.go
@@ -78,7 +78,7 @@ func (i *Item) addInt(x int64) {
 	i.numerator = i.numerator + addNum
 
 	// C.printf("addNum = %ld, numerator/denominator = %ld\n", add_num, numerator/denominator);
-	i.sum = float64(i.numerator / i.denominator)
+	i.sum = float64(i.numerator) / float64(i.denominator)
 
 	// i.sum = i.sum + x
 }
@@ -87,43 +87,26 @@ func (i *Item) subInt(x int64) {
 
 	// C.printf("Test add function\n");
 	subNum := x * i.denominator
-
 	i.numerator = i.numerator - subNum
-
 	// C.printf("subNum = %ld, i.numerator/i.denominator = %ld\n", subNum, i.numerator/i.denominator);
-	i.sum = float64(i.numerator / i.denominator)
+	i.sum = float64(i.numerator) / float64(i.denominator)
 
 	// i.sum = i.sum + x
 }
 
 func (i *Item) addFrac(num int64, den int64) {
 
-	// #if DEBUG_
-	// if den == 0 {
-	// 	 C.printf("WARNING: add_frac: den = 0\n")
-	// }
-	// if i.denominator == 0 {
-	//	 C.printf("WARNING: add_frac: 1. denominator = 0\n")
-	// }
-	// #endif
 
-	if i.denominator%den == 0 {
-		i.numerator = i.numerator + num*i.denominator/den
-	} else if den%i.denominator == 0 {
-		i.numerator = i.numerator*den/i.denominator + num
+	if i.denominator % den == 0 {
+		i.numerator = i.numerator + num * i.denominator / den
+	} else if den % i.denominator == 0 {
+		i.numerator = i.numerator * den / i.denominator + num
 		i.denominator = den
 	} else {
-		i.numerator = i.numerator*den + num*i.denominator
+		i.numerator = i.numerator * den + num * i.denominator
 		i.denominator = i.denominator * den
 	}
-
-	// #if DEBUG_
-	// if i.denominator == 0 {
-	//   C.printf("WARNING: addFrac: 2. denominator = 0\n")
-	// }
-	// #endif
-
-	i.sum = float64(i.numerator / i.denominator)
+	i.sum = float64(i.numerator) / float64(i.denominator)
 }
 
 func (i *Item) subFrac(num, den int64) {
@@ -145,48 +128,28 @@ func (i *Item) subFrac(num, den int64) {
 		i.denominator = i.denominator * den
 	}
 
-	// #if DEBUG_
-	// if denominator == 0
-	//	 C.printf("WARNING: subFrac: 2. denominator = 0\n")
-	// #endif
-
-	i.sum = float64(i.numerator / i.denominator)
+	i.sum = float64(i.numerator) / float64(i.denominator)
 }
 
 func (i *Item) demote() {
 	i.exponent = i.exponent + 1
 	den := int64(math.Exp2(i.exponent))
 	// C.printf("denominator = %ld\n", den);
-
 	i.subFrac(1, den)
 }
 
 func (i *Item) promote() {
 	den := int64(math.Exp2(i.exponent))
 
-	// #if DEBUG_
-	// if den == 0
-	//   C.printf("2 ^ %f = %ld?\n", i.exponent, den);
-	// #endif
-
 	if i.exponent < 0 {
 		den = 1
 	}
-
-	// C.printf("denominator = %ld\n", den);
-
 	i.addFrac(1, den)
+
 	i.exponent = i.exponent - 1
 }
 
 func (i *Item) addFracFailed(num int64, den int64) {
-
-	// #if DEBUG_
-	// if den == 0
-	//   C.printf("WARNING: addFracFailed: den = 0\n");
-	// if denominatorF == 0
-	//   C.printf("WARNING: addFracFailed: 1. denominatorF = 0\n")
-	// #endif
 
 	if i.denominatorF%den == 0 {
 		i.numeratorF = i.numeratorF + num*i.denominatorF/den
@@ -197,11 +160,6 @@ func (i *Item) addFracFailed(num int64, den int64) {
 		i.numeratorF = i.numeratorF*den + num*i.denominatorF
 		i.denominatorF = i.denominatorF * den
 	}
-
-	// #if DEBUG_
-	// if denominatorF == 0
-	//   C.printf("WARNING: addFracFailed: 2. denominatorF = 0\n");
-	// #endif
 
 	i.sumF = float64(i.numeratorF) / float64(i.denominatorF)
 }
@@ -255,7 +213,7 @@ func (i *Item) addFracReader(num int64, den int64) {
 		i.denominatorR = i.denominatorR * den
 	}
 
-	i.sumR = float64(i.numeratorR / i.denominatorR)
+	i.sumR = float64(i.numeratorR) / float64(i.denominatorR)
 }
 
 func (i *Item) subFracReader(num int64, den int64) {
@@ -269,7 +227,7 @@ func (i *Item) subFracReader(num int64, den int64) {
 		i.denominatorR = i.denominatorR * den
 	}
 
-	i.sumR = float64(i.numeratorR / i.denominatorR)
+	i.sumR = float64(i.numeratorR) / float64(i.denominatorR)
 }
 
 func (i *Item) demoteReader() {
